+++ conans/client/downloaders/file_downloader.py
@@ -7,8 +7,7 @@ from conan.api.output import ConanOutput, TimedOutput
 from conans.client.rest import response_to_str
 from conans.errors import ConanException, NotFoundException, AuthenticationException, \
     ForbiddenException, ConanConnectionError, RequestErrorException
-from conans.util.files import human_size
-from conans.util.sha import check_with_algorithm_sum
+from conans.util.files import human_size, check_with_algorithm_sum
 
 
 class FileDownloader:
