+++ conans/util/config_parser.py
@@ -3,24 +3,6 @@ import re
 from conans.errors import ConanException
 
 
-def get_bool_from_text_value(value):
-    """ to be deprecated
-    It has issues, as accepting into the registry whatever=value, as False, without
-    complaining
-    """
-    return (value == "1" or value.lower() == "yes" or value.lower() == "y" or
-            value.lower() == "true") if value else True
-
-
-def get_bool_from_text(value):
-    value = value.lower()
-    if value in ["1", "yes", "y", "true"]:
-        return True
-    if value in ["0", "no", "n", "false"]:
-        return False
-    raise ConanException("Unrecognized boolean value '%s'" % value)
-
-
 class ConfigParser(object):
     """ util class to load a file with sections as [section1]
     checking the values of those sections, and returns each section
